backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
python-nostr==0.3.0
websockets==12.0
sqlalchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4
bitcoin==1.1.42
lightning==0.1.0
web3==6.12.0
